[#24971] refactor(JmsIO): apply PR remarks #24973
Fixes #24971
